Add permission specifications and accept/reject hooks
Adds permission specifications, in the vein of caveat specifications, to
allow consumer-specified permission factory and validator functions. Now
that we have those, we also call caveat and permission validators (if
they exist) whenever a caveat and therefore permission is mutated, or
when either is instantiated.

Also adds methos for accepting and rejecting permission requests to the
permission enforcer.